+++ MbyronModsCommon/MbyronModsCommon/MbyronModsCommon.projitems
@@ -15,8 +15,6 @@
     <Compile Include="$(MSBuildThisFileDirectory)MessageBox\ResetModMessageBox.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)MessageBox\SimpleMessageBox.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)ModBase\ModBase.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)ModBase\ModConfigBase.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)ModBase\ModThreadExtensionBase.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)ModUtils\AssemblyUtils.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)ModUtils\CompatibilityCheck.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)ModUtils\DebugUtils.cs" />
